+++ Source/WebKit/WebProcess/GPU/graphics/ImageBufferShareableBitmapBackend.h
@@ -27,8 +27,13 @@
 
 #include "ImageBufferBackendHandleSharing.h"
 #include <WebCore/ImageBuffer.h>
-#include <WebCore/PlatformImageBufferBackend.h>
 #include <wtf/IsoMalloc.h>
+
+#if USE(CG)
+#include <WebCore/ImageBufferCGBackend.h>
+#elif USE(CAIRO)
+#include <WebCore/ImageBufferCairoBackend.h>
+#endif
 
 namespace WebCore {
 class ProcessIdentity;
@@ -38,7 +43,13 @@ namespace WebKit {
 
 class ShareableBitmap;
 
-class ImageBufferShareableBitmapBackend final : public WebCore::PlatformImageBufferBackend, public ImageBufferBackendHandleSharing {
+#if USE(CG)
+using ImageBufferShareableBitmapBackendBase = WebCore::ImageBufferCGBackend;
+#elif USE(CAIRO)
+using ImageBufferShareableBitmapBackendBase = WebCore::ImageBufferCairoBackend;
+#endif
+
+class ImageBufferShareableBitmapBackend final : public ImageBufferShareableBitmapBackendBase, public ImageBufferBackendHandleSharing {
     WTF_MAKE_ISO_ALLOCATED(ImageBufferShareableBitmapBackend);
     WTF_MAKE_NONCOPYABLE(ImageBufferShareableBitmapBackend);
